+++ apps/api/src/lib/extract/extraction-f1.ts
