+++ apps/api/src/services/extract-worker.ts
@@ -5,9 +5,9 @@ import { setSentryServiceTag } from "./sentry";
 import * as Sentry from "@sentry/node";
 import { logger as _logger } from "../lib/logger";
 import { configDotenv } from "dotenv";
-import { ExtractResult } from "../lib/extract/extraction-service";
+import { ExtractResult } from "../lib/extract/extraction-f1";
 import { updateExtract } from "../lib/extract/extract-redis";
-import { performExtraction_F0 } from "../lib/extract/fire-0/extraction-service-f0";
+import { performExtraction_F0 } from "../lib/extract/extraction-f0";
 import { createWebhookSender, WebhookEvent } from "./webhook";
 import Express from "express";
 import { getErrorContactMessage } from "../lib/deployment";
